+++ packages/@vuepress/core/__tests__/page/resolvePageComponentInfo.spec.ts
@@ -10,7 +10,6 @@ describe('core > page > resolvePageComponentInfo', () => {
   it('should resolve page component info correctly', async () => {
     const resolved = await resolvePageComponentInfo({
       app,
-      renderedContent: '<p>foobar</p>\n',
       hoistedTags: [],
       htmlFilePathRelative: 'foo.html',
       key: 'key',
@@ -19,7 +18,6 @@ describe('core > page > resolvePageComponentInfo', () => {
     expect(resolved).toEqual({
       componentFilePath: app.dir.temp('pages/foo.html.vue'),
       componentFilePathRelative: 'pages/foo.html.vue',
-      componentFileContent: '<template><p>foobar</p>\n</template>',
       componentFileChunkName: 'key',
     })
   })